requirements.txt
flet
python-dotenv
numpy
scikit-learn
paho-mqtt
